test merge

+++ src/main/java/com/webphim/webphim/Reponsitory/EpisodesRepository.java
@@ -3,5 +3,7 @@ package com.webphim.webphim.Reponsitory;
 import com.webphim.webphim.Model.Episodes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EpisodesRepository extends JpaRepository<Episodes, Integer> {
+import java.util.List;
+
+public interface EpisodesRepository extends JpaRepository<Episodes, Long> {
 }
