+++ src/main/java/com/webphim/webphim/Reponsitory/MoviesRepository.java
@@ -3,5 +3,5 @@ package com.webphim.webphim.Reponsitory;
 import com.webphim.webphim.Model.Movies;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoviesRepository extends JpaRepository<Movies, Integer> {
+public interface MoviesRepository extends JpaRepository<Movies, Long> {
 }
